moving funcs to files
